NCI-Agency/anet#387: Move email validation code to utils
