+++ client/src/utils.js
@@ -1,5 +1,10 @@
+import React from 'react'
 import * as changeCase from 'change-case'
 import pluralize from 'pluralize'
+
+import Settings from 'Settings'
+
+const WILDCARD = '*'
 
 export default {
 	...changeCase,
@@ -16,6 +21,64 @@ export default {
 			str += key + '=' + encodeURIComponent(val)
 		})
 		return str
+	},
+
+	handleEmailValidation: function(value, shouldValidate) {
+		const domainNames = Settings.domainNames.map(d => d.toLowerCase())
+		if (!shouldValidate || domainNames.length === 0) {
+			return { isValid: null, message: 'No custom validator is set' }
+		}
+
+		let wildcardDomains = this.getWildcardDomains(domainNames, WILDCARD)
+		let isValid = this.validateEmail(value, domainNames, wildcardDomains)
+
+		return { isValid: isValid, message: this.emailErrorMessage(domainNames) }
+	},
+
+	validateEmail: function(emailValue, domainNames, wildcardDomains) {
+		let email = emailValue.split('@')
+		let from =  email[0].trim()
+		let domain = email[1].toLowerCase()
+		return (
+			this.validateWithWhitelist(from, domain, domainNames) ||
+			this.validateWithWildcard(domain, wildcardDomains)
+		)
+	},
+
+	validateWithWhitelist: function(from, domain, whitelist) {
+		return from.length > 0 && whitelist.includes(domain)
+	},
+
+	validateWithWildcard: function(domain, wildcardDomains) {
+		let isValid = false
+		if (domain) {
+			isValid = wildcardDomains.some(wildcard => {
+				return domain[0] !== '.' && domain.endsWith(wildcard.substr(1))
+			})
+		}
+		return isValid
+	},
+
+	getWildcardDomains: function(domainList, token) {
+		let wildcardDomains = domainList.filter(domain => {
+			return domain[0] === token
+		})
+		return wildcardDomains
+	},
+
+	emailErrorMessage: function(validDomainNames) {
+		const supportEmail = Settings.SUPPORT_EMAIL_ADDR
+		const emailMessage = supportEmail ? ` at ${supportEmail}`: ''
+		const errorMessage = `Only the following email domain names are allowed. If your email domain name is not in the list, please contact the support team${emailMessage}.`
+		const items = validDomainNames.map(name => [
+			<li>{name}</li>
+		])
+		return (
+			<div>
+				<p>{errorMessage}</p>
+				<ul>{items}</ul>
+			</div>
+		)
 	}
 }
